[FEAT]: 회원탈퇴 기능 새로 구현 -> 모든 collection 삭제

+++ CarLog/CarLog/Models/Car.swift
@@ -13,6 +13,6 @@ struct Car: Codable {
     let name: String?
     let oilType: String?
     let nickName: String?
-    let totalDistance: Double?
+    let totalDistance: Int?
     let userEmail: String?
 }
